lint: staticcheck false positive fix
staticcheck seems to trigger false positives when user os.Exit(), it's
described in their docs as:

Staticcheck tries to deduce which functions abort control flow. For
example, it is aware that a function will not continue execution after a
call to panic or log.Fatal. However, sometimes this detection fails, in
particular in the presence of conditionals. Consider the following
example:

```
func Log(msg string, level int) {
        fmt.Println(msg)
            if level == levelFatal {
                        os.Exit(1)

            }

}

func Fatal(msg string) {
        Log(msg, levelFatal)

}

func fn(x *int) {
    if x == nil {
                Fatal("unexpected nil pointer")

    }
        fmt.Println(*x)

}
```

Staticcheck will flag the dereference of x, even though it is perfectly
safe. Staticcheck is not able to deduce that a call to Fatal will exit
the program. For the time being, the easiest workaround is to modify the
definition of Fatal like so:

```
func Fatal(msg string) {
        Log(msg, levelFatal)
            panic("unreachable")

}
```

Full doc: https://staticcheck.io/docs/checks#SA5011